Changes to run on local system

+++ dataset/data_loader/BaseLoader.py
@@ -11,7 +11,6 @@ import os
 import re
 from math import ceil
 from scipy import signal
-from scipy import sparse
 from unsupervised_methods.methods import POS_WANG
 from unsupervised_methods import utils
 import math
@@ -344,7 +343,7 @@ class BaseLoader(Dataset):
                     face_region = face_region_all[reference_index]
                 frame = frame[max(face_region[1], 0):min(face_region[1] + face_region[3], frame.shape[0]),
                         max(face_region[0], 0):min(face_region[0] + face_region[2], frame.shape[1])]
-            resized_frames[i] = cv2.resize(frame, (width, height), interpolation=cv2.INTER_AREA)
+                resized_frames[i] = cv2.resize(frame, (width, height), interpolation=cv2.INTER_AREA).astype('uint8')
         return resized_frames
 
     def chunk(self, frames, bvps, chunk_length):
@@ -587,7 +586,8 @@ class BaseLoader(Dataset):
                 1, input_signal.shape[0], target_length), np.linspace(
                 1, input_signal.shape[0], input_signal.shape[0]), input_signal)
 
-   @staticmethod
+
+    @staticmethod
     def resample_video(input_frames, target_length):
         T, H, W, C = input_frames.shape
         channel_data = input_frames.reshape(T, -1)
